added task auth and added appropriate policies

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -6,14 +6,19 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
+use Illuminate\Http\Request;
 
 class TaskController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Task::class)) {
+            abort(403);
+        }
+
         return request()
             ->user()
             ->tasks()
@@ -28,6 +33,9 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if ($request->user()->cannot('create', Task::class)) {
+            abort(403);
+        }
         $task = $request->user()->tasks()->create($request->validated());
 
         return $task->toResource();
@@ -36,8 +44,11 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
+        if ($request->user()->cannot('view', $task)) {
+            abort(403);
+        }
         return $task->toResource();
     }
 
@@ -46,16 +57,23 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403);
+        }
+        
         $task->update($request->validated());
-
+    
         return $task->toResource();
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        if ($request->user()->cannot('delete', $task)) {
+            abort(403);
+        }
         $task->delete();
 
         return response()->noContent();
